Add Windows system certificate pool support
This change copies the currently unused loadSystemRoots function from
crypto/x509/root_windows.go. Apparently, there are issues where this set
of root certificates is incomplete, with Windows downloading additional
certificates on demand. This change attempts to improve the situation
for many scenarios, while not addressing that problem.

Longer term, rather than trying to load all the system root
certificates, it might make sense to add a flag to verify entirely with
the Windows verification code. The flag would indicate that the
certificates specified in Roots be added to the in-memory context before
verifying the certificate.

Signed-off-by: John Stephens <johnstep@docker.com>